compute distance

+++ adam/Menu.swift
@@ -12,11 +12,11 @@ class Menu {
     var menuName: String   = "name"
     var storeName: String  = "res_name"
     var imgURL: NSURL      = NSURL(string: "http://someurl.com")!
-    var distanceVal:Float  = 1.0
+    var distanceVal:Double  = 1.0
     var pointVal:Int       = 1
     var price:Float        = 800
     
-    init(menuName: String, storeName: String, imgURL: NSURL, distanceVal: Float, pointVal: Int, price: Float) {
+    init(menuName: String, storeName: String, imgURL: NSURL, distanceVal: Double, pointVal: Int, price: Float) {
         self.menuName       = menuName
         self.storeName      = storeName
         self.imgURL         = imgURL
